Added support for using fields as the ID.
Also moved the properties into their own module ala Django.
